requirements.txt
Django==1.10.5
mysqlclient==1.3.9
django-debug-toolbar==1.6
pillow==4.2.1
ipython==5.5.0
django-rest-framework==0.1.0
social-auth-app-django==1.2.0

